+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.3.0
+- Continued work on `/src/MQFOverview`
+- Moved sidebar navigation to own Component
+- @todo: Add Meta viewport to `/public/index.html` to avoid inital scale on mobile
+- Commit `/public/`
+- Update `package.json` and `package-lock.json`
+
 ## 0.2.0
 - Updated mock state for handling authentication
 - Created dummy pages for handling the creation, editing, managing, studying, and testing of MQFs